Made a hook for the suggestCards function
