+++ components/Search/index.tsx
@@ -14,6 +14,7 @@ import {
 } from '@chakra-ui/icons'
 import type { LegalCards, ListAsProps } from '@/utils/dataTypes'
 import { useIsLegal } from '@/hooks/useIsLegal'
+import { useSuggestCards } from '@/hooks/useSuggestCards'
 
 interface Props {
   legalcards: LegalCards
@@ -26,6 +27,7 @@ const Search: FC<Props> = (props) => {
   const [cardIsLegal, setCardIsLegal] = useState(false)
   const [suggestions, setSuggestions] = useState<string[]>([])
   const { isLegal } = useIsLegal(legalCards)
+  const { suggestCards } = useSuggestCards(legalCards)
 
   const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     setExactMatch('')
@@ -42,33 +44,6 @@ const Search: FC<Props> = (props) => {
       setExactMatch(isLegalRet.exactMatch ?? '')
     }
     setSuggestions(suggestCards(newSearchBox))
-  }
-
-  const findOccurrences = (
-    needleTerm: string,
-    hayObject: ListAsProps,
-    limit: number = 40
-  ) => {
-    let occurrences: string[] = []
-    const keys = Object.keys(hayObject)
-    for (let i = 0; i < keys.length; i++) {
-      const key: string = keys[i]
-      let cardWeAreChecking = hayObject[key]
-      if (cardWeAreChecking === null) {
-        continue
-      }
-      if (cardWeAreChecking.toLowerCase().indexOf(needleTerm) != -1) {
-        occurrences.push(cardWeAreChecking)
-        if (occurrences.length >= limit) break
-      }
-    }
-    return occurrences
-  }
-
-  const suggestCards = (searchTerm: string, limit: number = 40) => {
-    let occurences = findOccurrences(searchTerm, legalCards.name, limit)
-    if (occurences.length > 0) return occurences
-    return findOccurrences(searchTerm, legalCards.name_ja, limit)
   }
 
   const placeholderIndex = ~~(
